Use one file per ingredient handler
