+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/data-constructor-properties.js
@@ -1,0 +1,10 @@
+const { linkListHandler } = require("./link-list-checker.js");
+
+/**
+ * Handler for the `data.constructor_properties` ingredient.
+ */
+const handleDataConstructorProperties = linkListHandler(
+  "Constructor_properties"
+);
+
+module.exports = handleDataConstructorProperties;
