+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/data-instance-methods.js
@@ -1,0 +1,8 @@
+const { linkListHandler } = require("./link-list-checker.js");
+
+/**
+ * Handler for the `data.instance_methods` ingredient.
+ */
+const handleDataInstanceMethods = linkListHandler("Instance_methods");
+
+module.exports = handleDataInstanceMethods;
