+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/data-instance-properties.js
@@ -1,0 +1,8 @@
+const { linkListHandler } = require("./link-list-checker.js");
+
+/**
+ * Handler for the `data.instance_properties` ingredient.
+ */
+const handleDataInstanceProperties = linkListHandler("Instance_properties");
+
+module.exports = handleDataInstanceProperties;
