+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/data-static-methods.js
@@ -1,0 +1,8 @@
+const { linkListHandler } = require("./link-list-checker.js");
+
+/**
+ * Handler for the `data.static_methods` ingredient.
+ */
+const handleDataStaticMethods = linkListHandler("Static_methods");
+
+module.exports = handleDataStaticMethods;
