+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/data-static-properties.js
@@ -1,0 +1,8 @@
+const { linkListHandler } = require("./link-list-checker.js");
+
+/**
+ * Handler for the `data.static_properties` ingredient.
+ */
+const handleDataStaticProperties = linkListHandler("Static_properties");
+
+module.exports = handleDataStaticProperties;
